app/course_stat_tool/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pyyaml>=6.0.0